refactor/extrai classe CodigoVoo para seu proprio arquivo, implementa teste gerarCodigo
